refactor(core): eliminate inverted core-to-feature imports and extract SSL and connection models to core
- Move ssl_certificate_support.dart to lib/core/security/
- Extract ConnectionType and ConnectionTypeChoice models to lib/core/database/
- Decouple ExtensionDriverCatalog and low-level DB drivers from UI features
- Eliminate circular imports between connection choice models and dialogs
- Add comprehensive unit tests in test/core/security/ and test/core/database/

Closes #644
