+++ lib/core/extensions/extension_driver_catalog.dart
@@ -1,9 +1,9 @@
+import 'package:querya_desktop/core/database/connection_type.dart';
+import 'package:querya_desktop/core/database/connection_type_choice.dart';
 import 'package:querya_desktop/core/extensions/local_extension_registry.dart';
 import 'package:querya_desktop/core/extensions/models/extension_manifest.dart';
 import 'package:querya_desktop/core/extensions/models/extension_type.dart';
 import 'package:querya_desktop/core/storage/local_db.dart';
-import 'package:querya_desktop/features/connections/connection_type_choice.dart';
-import 'package:querya_desktop/features/connections/new_connection_dialog.dart';
 
 /// Built-in + installed extension drivers for New Connection / Driver Manager.
 class ExtensionDriverCatalog {
